CI: add CI watchdog
This tries to generalize what an optional watchdog can be like,
obviously we're explicitly defining the variables but it does beg
the question if we should just keep on extending the .ci/ directory
with now an optional watchdog per repo. If so I guess we need just
one:

.ci/watchdog/guard --> the guard file
.ci/watchdog/run   --> the actual CI watchdog to run

The timeout can be shared between all CIs. If one is not defined
then clearly we'd skip this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>